# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "Answer": {
        "description": "What came back, and what it cost.\n\n`answered` is false when a reply was produced and then discarded — by the\noutput filters, or by the provider declining. The turn is spent, because the\nplayer had their go.\n\nAn unreachable model is not this shape: it is a 503 and costs nothing.\n\n`reason` stays coarse on purpose. The detail can quote the canary token the\nreply was discarded for, which is precisely the thing that must not travel.",
        "properties": {
          "alibi_broken": {
            "default": false,
            "description": "The evidence caught them out. A game event, not a fault (RN-021).",
            "title": "Alibi Broken",
            "type": "boolean"
          },
          "answered": {
            "title": "Answered",
            "type": "boolean"
          },
          "character": {
            "title": "Character",
            "type": "string"
          },
          "line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "`rejected` or `model_unavailable`, when unanswered.",
            "title": "Reason"
          },
          "stance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Stance"
              },
              {
                "type": "null"
              }
            ]
          },
          "turns_left": {
            "title": "Turns Left",
            "type": "integer"
          }
        },
        "required": [
          "answered",
          "character",
          "turns_left"
        ],
        "title": "Answer",
        "type": "object"
      },
      "CastMember": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "stance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Stance"
              },
              {
                "type": "null"
              }
            ],
            "description": "How they are holding up. Absent until they have been asked."
          }
        },
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "CastMember",
        "type": "object"
      },
      "Confrontation": {
        "properties": {
          "evidence": {
            "description": "Fact id the player holds, e.g. F-014.",
            "title": "Evidence",
            "type": "string"
          },
          "suspect": {
            "description": "Character id, e.g. sus-1.",
            "title": "Suspect",
            "type": "string"
          }
        },
        "required": [
          "suspect",
          "evidence"
        ],
        "title": "Confrontation",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Health": {
        "description": "Liveness probe response.",
        "properties": {
          "environment": {
            "enum": [
              "dev",
              "staging",
              "prod"
            ],
            "title": "Environment",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "environment"
        ],
        "title": "Health",
        "type": "object"
      },
      "KnownFact": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text"
        ],
        "title": "KnownFact",
        "type": "object"
      },
      "MatchState": {
        "description": "Everything a player may see. No solution, no bookkeeping.",
        "properties": {
          "cast": {
            "items": {
              "$ref": "#/components/schemas/CastMember"
            },
            "title": "Cast",
            "type": "array"
          },
          "evidence": {
            "default": [],
            "description": "Fact ids the player holds and may present.",
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "known": {
            "items": {
              "$ref": "#/components/schemas/KnownFact"
            },
            "title": "Known",
            "type": "array"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "notebook": {
            "items": {
              "$ref": "#/components/schemas/Said"
            },
            "title": "Notebook",
            "type": "array"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "turns_left": {
            "title": "Turns Left",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "seed",
          "locale",
          "turns_left",
          "cast",
          "known",
          "notebook"
        ],
        "title": "MatchState",
        "type": "object"
      },
      "NewAccusation": {
        "description": "One per match, irreversible (RN-031). Choose carefully.",
        "properties": {
          "culprit": {
            "description": "Character id you are naming.",
            "title": "Culprit",
            "type": "string"
          },
          "evidence": {
            "default": [],
            "description": "Fact ids you offer in support. Only ones you hold count.",
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          }
        },
        "required": [
          "culprit"
        ],
        "title": "NewAccusation",
        "type": "object"
      },
      "NewMatch": {
        "properties": {
          "locale": {
            "default": "pt-BR",
            "description": "Fixed for the match (ADR-0005).",
            "title": "Locale",
            "type": "string"
          },
          "seed": {
            "description": "Same seed, same mystery.",
            "title": "Seed",
            "type": "integer"
          },
          "suspects": {
            "default": 6,
            "maximum": 8.0,
            "minimum": 3.0,
            "title": "Suspects",
            "type": "integer"
          }
        },
        "required": [
          "seed"
        ],
        "title": "NewMatch",
        "type": "object"
      },
      "Outcome": {
        "description": "The verdict, and the only moment the solution may cross the wire.\n\nRN-011 governs a match in progress. This is its ending: the player has\nspent their one accusation and is owed the answer, right or wrong.",
        "properties": {
          "accused": {
            "title": "Accused",
            "type": "string"
          },
          "correct": {
            "title": "Correct",
            "type": "boolean"
          },
          "culprit": {
            "title": "Culprit",
            "type": "string"
          },
          "culprit_points": {
            "title": "Culprit Points",
            "type": "integer"
          },
          "evidence_expected": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Expected",
            "type": "array"
          },
          "evidence_hit": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Hit",
            "type": "array"
          },
          "evidence_points": {
            "title": "Evidence Points",
            "type": "integer"
          },
          "means": {
            "title": "Means",
            "type": "string"
          },
          "motive": {
            "title": "Motive",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "speed_points": {
            "title": "Speed Points",
            "type": "integer"
          },
          "turns_left": {
            "title": "Turns Left",
            "type": "integer"
          }
        },
        "required": [
          "correct",
          "accused",
          "culprit",
          "means",
          "motive",
          "score",
          "culprit_points",
          "evidence_points",
          "speed_points",
          "evidence_expected",
          "evidence_hit",
          "turns_left"
        ],
        "title": "Outcome",
        "type": "object"
      },
      "Question": {
        "properties": {
          "question": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "suspect": {
            "description": "Character id, e.g. sus-1.",
            "title": "Suspect",
            "type": "string"
          }
        },
        "required": [
          "suspect",
          "question"
        ],
        "title": "Question",
        "type": "object"
      },
      "Said": {
        "description": "One line in the notebook, as the player recorded it.",
        "properties": {
          "character": {
            "title": "Character",
            "type": "string"
          },
          "line": {
            "title": "Line",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "stance": {
            "$ref": "#/components/schemas/Stance"
          },
          "turn": {
            "title": "Turn",
            "type": "integer"
          }
        },
        "required": [
          "turn",
          "character",
          "question",
          "line",
          "stance"
        ],
        "title": "Said",
        "type": "object"
      },
      "Stance": {
        "description": "How a suspect is holding up. Transitions are code, never the model's call.\n\nThe model may suggest one; the backend validates it against the machine in\n`firenze.interrogation.stance` and keeps the current stance if the\nsuggestion is not a legal move (RN-023). `broken` is absorbing and is only\nreachable by confrontation, which arrives in phase 4.",
        "enum": [
          "cooperative",
          "evasive",
          "hostile",
          "broken"
        ],
        "title": "Stance",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Owns the whole deterministic domain and every call to a model (ADR-0003). The front end never talks to the model provider or to the database.",
    "summary": "AI core and domain of the mystery game.",
    "title": "Firenze API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "description": "Whether the process is up. Deliberately does not touch the database:\nreadiness with dependencies arrives with the schema (phase 1).",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness",
        "tags": [
          "infra"
        ]
      }
    },
    "/matches": {
      "post": {
        "operationId": "create_matches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start a match",
        "tags": [
          "match"
        ]
      }
    },
    "/matches/{match_id}": {
      "get": {
        "operationId": "read_matches__match_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The notebook",
        "tags": [
          "match"
        ]
      }
    },
    "/matches/{match_id}/accusation": {
      "post": {
        "description": "No model is involved. The outcome is arithmetic (RN-032).",
        "operationId": "accuse_matches__match_id__accusation_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewAccusation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Outcome"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Name a culprit and end the match",
        "tags": [
          "match"
        ]
      }
    },
    "/matches/{match_id}/confrontations": {
      "post": {
        "description": "Costs two turns, and only the evidence the player already holds.",
        "operationId": "take_confrontation_matches__match_id__confrontations_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Confrontation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show a suspect a piece of evidence",
        "tags": [
          "match"
        ]
      }
    },
    "/matches/{match_id}/turns": {
      "post": {
        "operationId": "take_turn_matches__match_id__turns_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Question"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Question a suspect",
        "tags": [
          "match"
        ]
      }
    }
  }
}